+++ src/main/java/jp/ne/divichart/DivichartApplication.java
@@ -1,4 +1,4 @@
-package com.example.divichart;
+package jp.ne.divichart;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
