+++ src/main/java/jp/ne/divichart/configuration/MvcConfig.java
@@ -1,4 +1,4 @@
-package com.example.divichart.configuration;
+package jp.ne.divichart.configuration;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
