+++ src/main/java/jp/ne/divichart/controller/PieChartController.java
@@ -1,6 +1,6 @@
-package com.example.divichart.controller;
+package jp.ne.divichart.controller;
 
-import com.example.divichart.service.BarChartService;
+import jp.ne.divichart.service.PieChartService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,18 +11,18 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/barChart")
-public class BarChartController {
+@RequestMapping("/pieChart")
+public class PieChartController {
 
     @Autowired
-    BarChartService service;
+    PieChartService service;
 
-    private static final Logger log = LoggerFactory.getLogger(BarChartController.class);
+    private static final Logger log = LoggerFactory.getLogger(PieChartController.class);
 
     @GetMapping
     public String index(Model model,
                         @ModelAttribute("targetYear") String targetYear) {
-        log.debug("月別配当グラフ表示");
+        log.debug("配当割合グラフ表示");
 
         String[] recentYears = service.getRecentYears();
         model.addAttribute("recentYears", recentYears);
@@ -30,10 +30,10 @@ public class BarChartController {
         if (targetYear.isEmpty()) targetYear = recentYears[0];
         model.addAttribute("targetYear", targetYear);
 
-        String chartData = service.getChartData(targetYear);
+        String[] chartData = service.getChartData(targetYear);
         model.addAttribute("chartData", chartData);
 
-        return "barChart";
+        return "pieChart";
     }
 
 }
