+++ src/main/java/jp/ne/divichart/service/BarChartService.java
@@ -1,4 +1,4 @@
-package com.example.divichart.service;
+package jp.ne.divichart.service;
 
 import org.springframework.stereotype.Service;
